pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "godd0t", email = "lirrishala@gmail.com"},
]
dependencies = [
    "superagi-tools>=1.0.6",
    "aiohttp>=3.8.5",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --log-cli-level=INFO"
testpaths = ["tests"]
pythonpath = ["."]
python_files = "tests.py test_*.py *_tests.py"
asyncio_mode = "auto"
